move combine errors into separate file
