+++ src/utils/time-entry-utils/timeEntryUtils.js
@@ -1,8 +1,6 @@
 import dayjs from 'dayjs';
 import isSameOrBefore from 'dayjs/plugin/isSameOrBefore';
-import TimeInputErrors from '../../components/timecard/time-input-errors/TimeInputErrors';
 import { deepCloneJson } from '../common-utils/common-utils';
-import { clashingProperties, inputNames } from '../constants';
 
 dayjs.extend(isSameOrBefore);
 
@@ -61,49 +59,3 @@ export const isFinishTimeOnNextDay = (startTimeValue, finishTimeValue) => {
   }
   return false;
 };
-
-export function combineExistingAndTimeClashErrors(
-  errors,
-  summaryErrors,
-  clashingProperty,
-  clashingTimes
-) {
-  const existingErrors =
-    Object.keys(errors).length > 0 ? errors : summaryErrors;
-
-  const combinedErrors = { ...existingErrors };
-
-  if (clashingProperty === clashingProperties.startAndEndTime) {
-    combinedErrors[inputNames.shiftStartTime] = {
-      message: (
-        <TimeInputErrors
-          clashingProperty={clashingProperty}
-          clashes={clashingTimes}
-        />
-      ),
-    };
-    combinedErrors[inputNames.shiftFinishTime] = {
-      message: '',
-    };
-  } else if (clashingProperty === clashingProperties.startTime) {
-    combinedErrors[inputNames.shiftStartTime] = {
-      message: (
-        <TimeInputErrors
-          clashingProperty={clashingProperty}
-          clashes={clashingTimes}
-        />
-      ),
-    };
-  } else if (clashingProperty === clashingProperties.endTime) {
-    combinedErrors[inputNames.shiftFinishTime] = {
-      message: (
-        <TimeInputErrors
-          clashingProperty={clashingProperty}
-          clashes={clashingTimes}
-        />
-      ),
-    };
-  }
-
-  return combinedErrors;
-}
